requirements.txt
# rental-mvp/requirements.txt
alembic==1.16.3
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2  
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing-extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
email-validator
pytest>=7.4
httpx>=0.27